Updates to ItemRecipe
Splitting the struct into three pieces, since there are redundant names from the JSON that we should be putting into arrays in order to manipulate later. We can now iterate across arrays to get information for the ith element. UNFINISHED: The array of objects is more complicated since it's an struct array, of an array. The regular reflect is unable to find the values inside the array.