Update xivapi.go
Used two maps to be able to find both the prices and the ingredient IDs with their elements common to each other. More efficient to use two maps, rather than mapping every single item to a price. This allows us to pick specific materials if they have recipes as well.